Move cloud sagas to dedi module
